requirements.txt
Flask
paho-mqtt
unicode_slugify
pathlib
